implemented without generics
